+++ android/app/src/main/java/com/teamtripdraw/android/data/httpClient/service/CreateRecordingPointService.kt
@@ -2,7 +2,7 @@ package com.teamtripdraw.android.data.httpClient.service
 
 import com.teamtripdraw.android.data.httpClient.dto.request.CreateRecordingPointRequest
 import com.teamtripdraw.android.data.httpClient.dto.response.CreateRecordingPointResponse
-import com.teamtripdraw.android.data.httpClient.retrofitAdapter.ResponseState
+import com.teamtripdraw.android.data.httpClient.retrofitAdapter.responseState.ResponseState
 import retrofit2.http.Body
 import retrofit2.http.POST
