+++ android/app/src/main/java/com/teamtripdraw/android/data/httpClient/service/DeletePointService.kt
@@ -1,6 +1,6 @@
 package com.teamtripdraw.android.data.httpClient.service
 
-import com.teamtripdraw.android.data.httpClient.retrofitAdapter.ResponseState
+import com.teamtripdraw.android.data.httpClient.retrofitAdapter.responseState.ResponseState
 import retrofit2.http.DELETE
 import retrofit2.http.Path
 import retrofit2.http.Query
