+++ android/app/src/main/java/com/teamtripdraw/android/data/httpClient/service/GetPointService.kt
@@ -1,7 +1,7 @@
 package com.teamtripdraw.android.data.httpClient.service
 
 import com.teamtripdraw.android.data.httpClient.dto.response.GetPointResponse
-import com.teamtripdraw.android.data.httpClient.retrofitAdapter.ResponseState
+import com.teamtripdraw.android.data.httpClient.retrofitAdapter.responseState.ResponseState
 import retrofit2.http.GET
 import retrofit2.http.Path
 import retrofit2.http.Query
