+++ android/app/src/main/java/com/teamtripdraw/android/data/httpClient/service/NicknameSetupService.kt
@@ -2,7 +2,7 @@ package com.teamtripdraw.android.data.httpClient.service
 
 import com.teamtripdraw.android.data.httpClient.dto.request.NicknameSetUpRequest
 import com.teamtripdraw.android.data.httpClient.dto.response.NicknameSetupResponse
-import com.teamtripdraw.android.data.httpClient.retrofitAdapter.ResponseState
+import com.teamtripdraw.android.data.httpClient.retrofitAdapter.responseState.ResponseState
 import retrofit2.http.Body
 import retrofit2.http.POST
