+++ android/app/src/main/java/com/teamtripdraw/android/data/httpClient/service/PostService.kt
@@ -3,7 +3,7 @@ package com.teamtripdraw.android.data.httpClient.service
 import com.teamtripdraw.android.data.httpClient.dto.response.AddPostResponse
 import com.teamtripdraw.android.data.httpClient.dto.response.GetPostListResponse
 import com.teamtripdraw.android.data.httpClient.dto.response.GetPostResponse
-import com.teamtripdraw.android.data.httpClient.retrofitAdapter.ResponseState
+import com.teamtripdraw.android.data.httpClient.retrofitAdapter.responseState.ResponseState
 import okhttp3.MultipartBody
 import okhttp3.RequestBody
 import retrofit2.http.DELETE
